# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API Dengue",
        "description": "Documentação da API em Laravel com Swagger",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000",
            "description": "Servidor local"
        }
    ],
    "paths": {
...
        "/api/fotos/{id}": {
            "get": {
                "tags": [
                    "Foto"
                ],
                "summary": "Exibe uma foto específica",
                "operationId": "9452aed03a104d1cf263da76c8efc8b5",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID da foto",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Foto retornada com sucesso"
                    },
                    "404": {
                        "description": "Foto não encontrada"
                    }
                }
            },
            "put": {
                "tags": [
                    "Foto"
                ],
                "summary": "Atualiza uma foto existente",
                "operationId": "71b915ba939322577bf56933d5bbb86e",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID da foto",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "properties": {
                                    "foto": {
                                        "description": "Arquivo da nova foto",
                                        "type": "string",
                                        "format": "binary"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Foto atualizada com sucesso"
                    },
                    "404": {
                        "description": "Foto não encontrada"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Foto"
                ],
                "summary": "Remove uma foto",
                "operationId": "ba54ad653491fe9f59294ee9ba81bc57",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID da foto",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Foto excluída com sucesso"
                    },
                    "404": {
                        "description": "Foto não encontrada"
                    }
                }
            }
        },
        "/api/fotos/{idDenuncia}": {
            "post": {
                "tags": [
                    "Foto"
                ],
                "summary": "Cadastra uma nova foto para uma denúncia",
                "operationId": "bc301d633a5d849a037998c18b0cfdc3",
                "parameters": [
                    {
                        "name": "idDenuncia",
                        "in": "path",
                        "description": "ID da denúncia",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "required": [
                                    "foto"
                                ],
                                "properties": {
                                    "foto": {
                                        "description": "Arquivo da foto",
                                        "type": "string",
                                        "format": "binary"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Foto salva com sucesso"
                    },
                    "400": {
                        "description": "Nenhuma foto enviada"
                    }
                }
            }
        },
...
    },
...
}